__ni_wchar_escape: explicitly cast wchar_t to int when passing it as argument to printf("\\%03o")

+++ dborb/util.c
@@ -1673,7 +1673,7 @@ __ni_wchar_escape(wchar_t wc, const char **retval)
 	}
 
 	for (i = slen = 0; i < nbytes; ++i) {
-		snprintf(escbuf + slen, sizeof(escbuf) - slen, "\\%03o", rev & 0xFF);
+		snprintf(escbuf + slen, sizeof(escbuf) - slen, "\\%03o", (unsigned int) rev & 0xFF);
 		rev >>= 8;
 		slen += 4;
 	}
